Set "useDefineForClassFields" of tsconfig.json to "false".
I hoped that set it to "true".
But when that, the webpack was not able to bundle them with the following error.
It explains the babel-loader is needed, I think.

```bash
ERROR in ./lib/storage/Mock.ts 2:8
Module parse failed: Unexpected token (2:8)
File was processed with these loaders:
 * ../node_modules/ts-loader/index.js
You may need an additional loader to handle the result of these loaders.
| class Storage {
>     data;
|     constructor() {
|         this.data = {};
 @ ./lib/storage/index.ts 3:0-26 9:36-40 18:23-27
 @ ./background/message/repository/counter/index.ts 1:0-40 10:16-23
 @ ./background/message/repository/index.ts 1:0-42 7:12-29
 @ ./background/message/index.ts 1:0-48 7:16-36
 @ ./background/index.ts 1:0-37 3:4-2
```